Add temperature switch function
