+++ src/styles.scss
@@ -208,7 +208,6 @@ h2 {
     box-shadow: 0 10px 10px rgba(#fe4551, 0.5);
   } 
 
-
   &-done {
     @extend .toggleWrapper;
